torrentscsv: bunt datestamp from int to long. resolves #6550

+++ src/Jackett.Common/Indexers/Torrentscsv.cs
@@ -137,7 +137,7 @@ namespace Jackett.Common.Indexers {
                     release.InfoHash = torrent.Value<JToken> ("infohash").ToString ();
 
                     // convert unix timestamp to human readable date
-                    double createdunix = torrent.Value<int> ("created_unix");
+                    double createdunix = torrent.Value<long> ("created_unix");
                     System.DateTime dateTime = new System.DateTime (1970, 1, 1, 0, 0, 0, 0);
                     dateTime = dateTime.AddSeconds (createdunix);
                     release.PublishDate = dateTime;
